Fix early exit interrupt bombing.
Previous TWI change introduced early exit from slave wait function, when
packet was received but stop condition is not yet received. Then pato
processes the packet while TWINT flagis left high keeping the CLK low.

Unfortunately this led to the TWI module to call the ISR again and again
(this is normal, according to the data-sheet). The fix turns additional
flag - TWIE till the skip or send operation to get rid of the interrupt bomb.