Fix tty_state.skip_til_nl surviving display reset.
The state variable was not reset by display resetting. This is wrong in scenario
when printing line without \n, resetting display and printing a line again.
The second line is not printed since skip_til_nl is set to true.

Introduced hd44780_print_reset() that is called on display reset.